Changes to build process so that multiple Kindle versions can be supported by one installer.
Devices list based on the devices used in NiLuJe's USBNetwork Hack, which was the newest released package at the time of this work. In addition to the devices now included in the Makefile, the USBNetwork build script also includes --device paperwhite 5 --device basic4 --device scribe in the "new" devices, which the latest released version of kidletool doesn't appear to support. Thus, I have left those commented out for now.